Former-commit-id: b5017601a78eb4ddd76e5ea4620396bfe58a57aa

+++ cave/com.raytheon.viz.mpe.ui/src/com/raytheon/viz/mpe/ui/actions/SaveBestEstimate.java
@@ -289,8 +289,10 @@ public class SaveBestEstimate {
 
         }
         File qpePolyFile = PolygonEditManager.getHourlyEditFile(bestEstField, editDate);
-        if(qpePolyFile.exists()){
-            qpePolyFile.delete();
+        if (bestEstField.getFieldName().equals(DisplayFieldData.Xmrg.getFieldName())){
+            if (qpePolyFile.exists()){
+                qpePolyFile.delete();
+            }
         }
         MPEDisplayManager.getCurrent().setSavedData(true);
     }
